polygon/polygon_server/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
python-dotenv
psutil
httpx